services/ misc clean up
