+++ src/services/marketListings/index.ts
@@ -1,4 +1,2 @@
 export * from './types.ts'
-export * from './actions.ts'
-export * from './view.ts'
 export * from './api.ts'
